move execution to fishnet-cod, bump to 0.0.7
