+++ examples/fishnet_executor/requirements.txt
@@ -1,4 +1,3 @@
 aleph-client
 aars
 fishnet-cod
-pandas
